requirements.txt
streamlit
pandas
numpy
seaborn
google-cloud-bigquery
google-auth-oauthlib
scipy
matplotlib
